Add checks for used label and device with active hold.
When creating a new hold, this adds a check for making sure the provided
label exists in the (fake) labels database.
A check is also made for making sure the provided label is not already
in use by other active holds. This is done because a label, meaning a
physical station, can not be used by two currently active holds.