Rework release hold endpoint.
This makes the ReleaseHold endpoint take a device id instead of a hold
id. Instead of change the "end" value of a specific hold it now
automatically find the lastest hold on the given device, and changes
that end value.
This makes misuse of the endpoint less likely, as you cannot change
previous hold's end values.