Implement reworked database model 3.0.
This makes the API work with the reworked datamodel version 3.0, which
should be able to handle real data from Cordulus' tycho system. This
means endpoints, DAO's and tests have been adjusted to work with the new
datamodel. The rework paves the way for migrating Cordulus' event data
into the system.