Add fake stations database.
This adds a fake implementations of a stations database which contains a
map of labels and their imeis. It replaces the previous fake labels
database but keep this previous functionality. The new added
functionality is for getting the imei of a provided label. Customers do
not know what the imei of their weather station is, they only know the
label, but the system needs to translate from label to imei, to be able
to get data for the station.