Add GetMeasurements endpoint and logic.
This adds an endpoint for getting a continuous data chain for a device.
All valid sensors except rain returns an average for each interval. Rain
provides a sum of all measurements inside each interval. The timestamp
of intervals is the end time, meaning the interval goes from last
intervals timestamp to the current's timestamp.