Add edit holds endpoint.
This adds an endpoint for editing a hold's start and end attributes. It
comes with a number a restraints which will uphold the integrity of the
given hold in relation to the system.
If the hold is active you cannot edit the end attribute, as this should
be done through releasing or replacing a hold.